fix：代码提交

+++ src/main/java/com/ujs/shop/service/impl/CategoryServiceImpl.java
@@ -1,5 +1,6 @@
 package com.ujs.shop.service.impl;
 
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
@@ -10,9 +11,13 @@ import com.ujs.shop.common.exception.ServiceException;
 import com.ujs.shop.common.global.ConstantBean;
 import com.ujs.shop.common.global.PageFormBean;
 import com.ujs.shop.common.po.CategoryPO;
+import com.ujs.shop.common.po.GoodsPO;
+import com.ujs.shop.common.po.PackagePO;
 import com.ujs.shop.common.ro.AddCategoryRO;
 import com.ujs.shop.common.ro.UpdateCategoryRO;
 import com.ujs.shop.mapper.CategoryMapper;
+import com.ujs.shop.mapper.GoodsMapper;
+import com.ujs.shop.mapper.PackageMapper;
 import com.ujs.shop.service.CategoryService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,6 +25,7 @@ import org.springframework.stereotype.Service;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.List;
 
 /**
  * @author mundo.wang
@@ -33,6 +39,12 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, CategoryPO>
 
     @Autowired
     private CategoryMapper categoryMapper;
+
+    @Autowired
+    private GoodsMapper goodsMapper;
+
+    @Autowired
+    private PackageMapper packageMapper;
 
     @Override
     public void addCategory(AddCategoryRO addCategoryRO) {
@@ -57,6 +69,15 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, CategoryPO>
         CategoryPO categoryPO = categoryMapper.selectById(id);
         if (categoryPO == null) {
             throw new ServiceException(ResponseCodeEnum.NO_SUCH_CATEGORY);
+        }
+        LambdaQueryWrapper<GoodsPO> wrapper1 = new LambdaQueryWrapper<>();
+        wrapper1.eq(GoodsPO::getCategoryId, id);
+        List<GoodsPO> goodsPOList = goodsMapper.selectList(wrapper1);
+        LambdaQueryWrapper<PackagePO> wrapper2 = new LambdaQueryWrapper<>();
+        wrapper2.eq(PackagePO::getCategoryId, id);
+        List<PackagePO> packagePOS = packageMapper.selectList(wrapper2);
+        if (goodsPOList.size() != 0 || packagePOS.size() != 0) {
+            throw new ServiceException(ResponseCodeEnum.HAS_GOODS_OR_PACKAGE);
         }
         categoryMapper.deleteById(id);
     }
